@@ -11,19 +11,19 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^20.3.16",
+    "@angular/animations": "^21.1.2",
     "@angular/cdk": "^20.2.14",
-    "@angular/common": "^20.3.16",
-    "@angular/compiler": "^20.3.16",
-    "@angular/core": "^20.3.16",
+    "@angular/common": "^21.1.2",
+    "@angular/compiler": "^21.1.2",
+    "@angular/core": "^21.1.2",
     "@angular/flex-layout": "^15.0.0-beta.42",
-    "@angular/forms": "^20.3.16",
+    "@angular/forms": "^21.1.2",
     "@angular/material": "^20.2.14",
-    "@angular/platform-browser": "^20.3.16",
-    "@angular/platform-browser-dynamic": "^20.3.16",
-    "@angular/platform-server": "^20.3.16",
-    "@angular/router": "^20.3.16",
-    "@angular/ssr": "^20.3.15",
+    "@angular/platform-browser": "^21.1.2",
+    "@angular/platform-browser-dynamic": "^21.1.2",
+    "@angular/platform-server": "^21.1.2",
+    "@angular/router": "^21.1.2",
+    "@angular/ssr": "^21.1.2",
     "express": "^4.18.2",
     "jwt-decode": "^4.0.0",
     "ngx-cookie-service": "^21.1.0",
@@ -32,9 +32,9 @@
     "zone.js": "~0.15.1"
   },
   "devDependencies": {
-    "@angular/build": "^20.3.15",
-    "@angular/cli": "^20.3.15",
-    "@angular/compiler-cli": "^20.3.16",
+    "@angular/build": "^21.1.2",
+    "@angular/cli": "^21.1.2",
+    "@angular/compiler-cli": "^21.1.2",
     "@types/express": "^4.17.17",
     "@types/jasmine": "~5.1.0",
     "@types/node": "^18.18.0",
