@@ -1,12 +1,12 @@
 import { Component } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
 import { LoginService } from './pages/login/services/login.service';
-import { CommonModule } from '@angular/common';
+
 import { ApiLoaderComponent } from './core/services/Api/api.service';
 
 @Component({
     selector: 'app-root',
-    imports: [RouterOutlet, CommonModule, ApiLoaderComponent],
+    imports: [RouterOutlet, ApiLoaderComponent],
     templateUrl: './app.component.html',
     styleUrl: './app.component.css'
 })
