@@ -13,7 +13,7 @@ import { MatSelectModule } from '@angular/material/select';
 import { Router, RouterLink } from '@angular/router';
 import { ClienteModel } from '../../../models/cliente.model';
 import { CidadeModel } from '../../../../../core/model/cidade.model';
-import { CommonModule } from '@angular/common';
+
 import { ClienteService } from '../../../services/cliente.service';
 import { SharedModule } from '../../../../../shared/components/Module/shared.module';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
