@@ -1,4 +1,4 @@
-import { CommonModule, DatePipe } from '@angular/common';
+import { DatePipe } from '@angular/common';
 import { Component, OnInit, ViewChild, AfterViewInit } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
 import { MatIcon, MatIconModule } from '@angular/material/icon';
@@ -24,13 +24,12 @@ import { GrupoFiltroComponent } from '../grupo-filtro/grupo-filtro.component';
 @Component({
     selector: 'app-grupo-list',
     imports: [
-        MatPaginatorModule,
-        MatTableModule,
-        MatButtonModule,
-        MatIcon,
-        MatIconModule,
-        CommonModule,
-    ],
+    MatPaginatorModule,
+    MatTableModule,
+    MatButtonModule,
+    MatIcon,
+    MatIconModule
+],
     providers: [
         {
             provide: MatPaginatorIntl,
